+++ ibis/backends/pyspark/__init__.py
@@ -1,5 +1,5 @@
-from ibis.pyspark.client import PySparkClient
-from ibis.pyspark.compiler import dialect  # noqa: F401
+from .client import PySparkClient
+from .compiler import dialect  # noqa: F401
 
 
 def connect(session):
