+++ ibis/backends/pyspark/client.py
@@ -2,11 +2,14 @@ from pyspark.sql.column import Column
 
 import ibis.common.exceptions as com
 import ibis.expr.types as types
+from ibis.backends.pyspark.compiler import (
+    PySparkDialect,
+    PySparkExprTranslator,
+)
+from ibis.backends.pyspark.operations import PySparkTable
 from ibis.backends.spark.client import SparkClient
 from ibis.expr.scope import Scope
 from ibis.expr.timecontext import canonicalize_context
-from ibis.pyspark.compiler import PySparkDialect, PySparkExprTranslator
-from ibis.pyspark.operations import PySparkTable
 
 
 class PySparkClient(SparkClient):
