+++ ibis/backends/pyspark/compiler.py
@@ -14,6 +14,11 @@ import ibis.expr.types as ir
 import ibis.expr.types as types
 from ibis import interval
 from ibis.backends.pandas.execution import execute
+from ibis.backends.pyspark.operations import PySparkTable
+from ibis.backends.pyspark.timecontext import (
+    combine_time_context,
+    filter_by_time_context,
+)
 from ibis.backends.spark.compiler import SparkContext, SparkDialect
 from ibis.backends.spark.datatypes import (
     ibis_array_dtype_to_spark_dtype,
@@ -21,11 +26,6 @@ from ibis.backends.spark.datatypes import (
     spark_dtype,
 )
 from ibis.expr.timecontext import adjust_context
-from ibis.pyspark.operations import PySparkTable
-from ibis.pyspark.timecontext import (
-    combine_time_context,
-    filter_by_time_context,
-)
 
 
 class PySparkContext(SparkContext):
